+++ sdk/netapp/Azure.ResourceManager.NetApp/tests/ResourceTests/ANFBackupTests.cs
@@ -18,15 +18,14 @@ namespace Azure.ResourceManager.NetApp.Tests
 {
     public class ANFBackupTests: NetAppTestBase
     {
-        private NetAppAccountCollection _netAppAccountCollection { get => _resourceGroup.GetNetAppAccounts(); }
+        private static NetAppAccountCollection _netAppAccountCollection { get => _resourceGroup.GetNetAppAccounts(); }
         private readonly string _pool1Name = "pool1";
-        public static new AzureLocation DefaultLocation = AzureLocation.WestUS2;
-        public static new AzureLocation DefaultLocationString = DefaultLocation;
+        private readonly AzureLocation _defaultLocation = AzureLocation.WestUS2;
+        //private new readonly AzureLocation _defaultLocationString = _defaultLocation;
         internal NetAppAccountBackupCollection _accountBackupCollection;
         internal NetAppVolumeBackupCollection _volumeBackupCollection;
         internal NetAppVolumeResource _volumeResource;
-        internal List<NetAppVault> _vaults;
-        internal NetAppVault _vault;
+
         public ANFBackupTests(bool isAsync) : base(isAsync)
         {
         }
@@ -34,22 +33,19 @@ namespace Azure.ResourceManager.NetApp.Tests
         [SetUp]
         public async Task SetUp()
         {
-            _resourceGroup = await CreateResourceGroupAsync(location:DefaultLocation);
-            string accountName = await CreateValidAccountNameAsync(_accountNamePrefix, _resourceGroup, DefaultLocation);
-            _netAppAccount = (await _netAppAccountCollection.CreateOrUpdateAsync(WaitUntil.Completed, accountName, GetDefaultNetAppAccountParameters(location:DefaultLocation))).Value;
+            _resourceGroup = await CreateResourceGroupAsync(location:_defaultLocation);
+            string accountName = await CreateValidAccountNameAsync(_accountNamePrefix, _resourceGroup, _defaultLocation);
+            _netAppAccount = (await _netAppAccountCollection.CreateOrUpdateAsync(WaitUntil.Completed, accountName, GetDefaultNetAppAccountParameters(location:_defaultLocation))).Value;
 
-            CapacityPoolData capactiyPoolData = new(DefaultLocation, _poolSize.Value, NetAppFileServiceLevel.Premium);
+            CapacityPoolData capactiyPoolData = new(_defaultLocation, _poolSize.Value, NetAppFileServiceLevel.Premium);
             capactiyPoolData.Tags.InitializeFrom(DefaultTags);
             _capacityPool = (await _capacityPoolCollection.CreateOrUpdateAsync(WaitUntil.Completed, _pool1Name, capactiyPoolData)).Value;
             _volumeCollection = _capacityPool.GetNetAppVolumes();
 
-            DefaultVirtualNetwork = await CreateVirtualNetwork(location:DefaultLocation);
-            _volumeResource = await CreateVolume(DefaultLocation, NetAppFileServiceLevel.Premium, _defaultUsageThreshold, subnetId: DefaultSubnetId);
+            DefaultVirtualNetwork = await CreateVirtualNetwork(location:_defaultLocation);
+            _volumeResource = await CreateVolume(_defaultLocation, NetAppFileServiceLevel.Premium, _defaultUsageThreshold, subnetId: DefaultSubnetId);
             _accountBackupCollection = _netAppAccount.GetNetAppAccountBackups();
             _volumeBackupCollection = _volumeResource.GetNetAppVolumeBackups();
-            //getVault id
-            _vaults = await _netAppAccount.GetVaultsAsync().ToEnumerableAsync();
-            _vault = _vaults.FirstOrDefault();
         }
 
         [TearDown]
@@ -58,7 +54,7 @@ namespace Azure.ResourceManager.NetApp.Tests
             //remove all volumes and backups under current capcityPool, remove pool and netAppAccount
             if (_resourceGroup != null && _capacityPoolCollection != null)
             {
-                bool exists = await _capacityPoolCollection.ExistsAsync(_capacityPool.Id.Name);
+                _ = await _capacityPoolCollection.ExistsAsync(_capacityPool.Id.Name);
                 CapacityPoolCollection poolCollection = _netAppAccount.GetCapacityPools();
                 List<CapacityPoolResource> poolList = await poolCollection.GetAllAsync().ToEnumerableAsync();
                 string lastBackupName = string.Empty;
@@ -87,11 +83,15 @@ namespace Azure.ResourceManager.NetApp.Tests
                         //disable backups if enabled
                         if (volume.Data.DataProtection?.Backup?.IsBackupEnabled == true)
                         {
-                            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, _vault.Id, false);
-                            NetAppVolumePatchDataProtection dataProtectionProperties = new();
-                            dataProtectionProperties.Backup = backupPolicyProperties;
-                            NetAppVolumePatch volumePatch = new(DefaultLocation);
-                            volumePatch.DataProtection = dataProtectionProperties;
+                            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, false);
+                            NetAppVolumePatchDataProtection dataProtectionProperties = new()
+                            {
+                                Backup = backupPolicyProperties
+                            };
+                            NetAppVolumePatch volumePatch = new(_defaultLocation)
+                            {
+                                DataProtection = dataProtectionProperties
+                            };
                             await _volumeResource.UpdateAsync(WaitUntil.Completed, volumePatch);
                         }
                         if (Mode != RecordedTestMode.Playback)
@@ -140,20 +140,22 @@ namespace Azure.ResourceManager.NetApp.Tests
         }
 
         [Test]
-        [Ignore("Ignore for now due to service side issue, re-enable when service side issue is fixed")]
+        //[Ignore("Ignore for now due to service side issue, re-enable when service side issue is fixed")]
         [RecordedTest]
         public async Task CreateDeleteBackup()
         {
             Console.WriteLine($"{DateTime.Now} Test CreateDeleteBackup");
-            _vaults.Should().HaveCount(1);
-            Assert.IsNotNull(_vault);
 
             //Update volume to enable backups
-            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, _vault.Id, true);
-            NetAppVolumePatchDataProtection dataProtectionProperties = new();
-            dataProtectionProperties.Backup = backupPolicyProperties;
-            NetAppVolumePatch volumePatch = new(DefaultLocation);
-            volumePatch.DataProtection = dataProtectionProperties;
+            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, true);
+            NetAppVolumePatchDataProtection dataProtectionProperties = new()
+            {
+                Backup = backupPolicyProperties
+            };
+            NetAppVolumePatch volumePatch = new(_defaultLocation)
+            {
+                DataProtection = dataProtectionProperties
+            };
             NetAppVolumeResource volumeResource1 = (await _volumeResource.UpdateAsync(WaitUntil.Completed, volumePatch)).Value;
             if (Mode != RecordedTestMode.Playback)
             {
@@ -169,8 +171,10 @@ namespace Azure.ResourceManager.NetApp.Tests
 
             //create Backup
             var backupName = Recording.GenerateAssetName("backup-");
-            NetAppBackupData backupData = new(DefaultLocation);
-            backupData.Label = "adHocBackup";
+            NetAppBackupData backupData = new(_defaultLocation)
+            {
+                Label = "adHocBackup"
+            };
             NetAppVolumeBackupResource backupResource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, backupName, backupData)).Value;
             Assert.IsNotNull(backupResource1);
             Assert.AreEqual(backupName, backupResource1.Id.Name);
@@ -207,7 +211,7 @@ namespace Azure.ResourceManager.NetApp.Tests
             //Delete backup
             //create another Backup
             var secondBackupName = Recording.GenerateAssetName("secondBackup-");
-            NetAppBackupData backupData2 = new(DefaultLocation);
+            NetAppBackupData backupData2 = new(_defaultLocation);
             backupData.Label = "secondAdHocBackup";
             NetAppVolumeBackupResource secondBackupResource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, secondBackupName, backupData2)).Value;
             Assert.IsNotNull(secondBackupResource1);
@@ -217,8 +221,9 @@ namespace Azure.ResourceManager.NetApp.Tests
                 await Task.Delay(60000);
             }
             await WaitForBackupSucceeded(_volumeBackupCollection, secondBackupName);
+
             //Test delete action on backup deleting the second backup
-            await secondBackupResource1.DeleteAsync(WaitUntil.Completed);
+            await backupResource2.DeleteAsync(WaitUntil.Completed);
             if (Mode != RecordedTestMode.Playback)
             {
                 await Task.Delay(40000);
@@ -257,16 +262,16 @@ namespace Azure.ResourceManager.NetApp.Tests
         [RecordedTest]
         public async Task UpdateBackup()
         {
-            //getVault id
-            _vaults.Should().HaveCount(1);
-            Assert.IsNotNull(_vault);
-
             //Update volume to enable backups
-            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, _vault.Id, true);
-            NetAppVolumePatchDataProtection dataProtectionProperties = new();
-            dataProtectionProperties.Backup = backupPolicyProperties;
-            NetAppVolumePatch volumePatch = new(DefaultLocation);
-            volumePatch.DataProtection = dataProtectionProperties;
+            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, true);
+            NetAppVolumePatchDataProtection dataProtectionProperties = new()
+            {
+                Backup = backupPolicyProperties
+            };
+            NetAppVolumePatch volumePatch = new(_defaultLocation)
+            {
+                DataProtection = dataProtectionProperties
+            };
             NetAppVolumeResource volumeResource1 = (await _volumeResource.UpdateAsync(WaitUntil.Completed, volumePatch)).Value;
             if (Mode != RecordedTestMode.Playback)
             {
@@ -282,8 +287,10 @@ namespace Azure.ResourceManager.NetApp.Tests
 
             //create Backup
             var backupName = Recording.GenerateAssetName("backup-");
-            NetAppBackupData backupData = new(DefaultLocation);
-            backupData.Label = "adHocBackup";
+            NetAppBackupData backupData = new(_defaultLocation)
+            {
+                Label = "adHocBackup"
+            };
             NetAppVolumeBackupResource backupResource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, backupName, backupData)).Value;
             Assert.IsNotNull(backupResource1);
             Assert.AreEqual(backupName, backupResource1.Id.Name);
@@ -299,8 +306,10 @@ namespace Azure.ResourceManager.NetApp.Tests
             Assert.IsFalse(await _volumeBackupCollection.ExistsAsync(backupName + "1"));
 
             //Update backup
-            NetAppVolumeBackupPatch backupPatch = new();
-            backupPatch.Label = "updatedLabel";
+            NetAppVolumeBackupPatch backupPatch = new()
+            {
+                Label = "updatedLabel"
+            };
             NetAppVolumeBackupResource backupResource3 = (await backupResource2.UpdateAsync(WaitUntil.Completed, backupPatch)).Value;
             await WaitForBackupSucceeded(_volumeBackupCollection, backupName);
             //Validate
@@ -315,16 +324,16 @@ namespace Azure.ResourceManager.NetApp.Tests
         [RecordedTest]
         public async Task ListBackups()
         {
-            //getVault id
-            _vaults.Should().HaveCount(1);
-            Assert.IsNotNull(_vault);
-
             //Update volume to enable backups
-            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, _vault.Id, true);
-            NetAppVolumePatchDataProtection dataProtectionProperties = new();
-            dataProtectionProperties.Backup = backupPolicyProperties;
-            NetAppVolumePatch volumePatch = new(DefaultLocation);
-            volumePatch.DataProtection = dataProtectionProperties;
+            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, true);
+            NetAppVolumePatchDataProtection dataProtectionProperties = new()
+            {
+                Backup = backupPolicyProperties
+            };
+            NetAppVolumePatch volumePatch = new(_defaultLocation)
+            {
+                DataProtection = dataProtectionProperties
+            };
             NetAppVolumeResource volumeResource1 = (await _volumeResource.UpdateAsync(WaitUntil.Completed, volumePatch)).Value;
             if (Mode != RecordedTestMode.Playback)
             {
@@ -340,8 +349,10 @@ namespace Azure.ResourceManager.NetApp.Tests
 
             //create Backup
             var backupName = Recording.GenerateAssetName("backup-");
-            NetAppBackupData backupData = new(DefaultLocation);
-            backupData.Label = "adHocBackup";
+            NetAppBackupData backupData = new(_defaultLocation)
+            {
+                Label = "adHocBackup"
+            };
             NetAppVolumeBackupResource backupResource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, backupName, backupData)).Value;
             Assert.IsNotNull(backupResource1);
             Assert.AreEqual(backupName, backupResource1.Id.Name);
@@ -354,7 +365,7 @@ namespace Azure.ResourceManager.NetApp.Tests
 
             //create second Backup
             var backupName2 = Recording.GenerateAssetName("backup-");
-            NetAppBackupData backupData2 = new(DefaultLocation);
+            NetAppBackupData backupData2 = new(_defaultLocation);
             backupData.Label = "adHocBackup2";
             NetAppVolumeBackupResource backup2Resource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, backupName2, backupData2)).Value;
             Assert.IsNotNull(backup2Resource1);
@@ -388,16 +399,16 @@ namespace Azure.ResourceManager.NetApp.Tests
         [RecordedTest]
         public async Task ListAccountBackups()
         {
-            //getVault id
-            _vaults.Should().HaveCount(1);
-            Assert.IsNotNull(_vault);
-
             //Update volume to enable backups
-            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, _vault.Id, true);
-            NetAppVolumePatchDataProtection dataProtectionProperties = new();
-            dataProtectionProperties.Backup = backupPolicyProperties;
-            NetAppVolumePatch volumePatch = new(DefaultLocation);
-            volumePatch.DataProtection = dataProtectionProperties;
+            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, true);
+            NetAppVolumePatchDataProtection dataProtectionProperties = new()
+            {
+                Backup = backupPolicyProperties
+            };
+            NetAppVolumePatch volumePatch = new(_defaultLocation)
+            {
+                DataProtection = dataProtectionProperties
+            };
             NetAppVolumeResource volumeResource1 = (await _volumeResource.UpdateAsync(WaitUntil.Completed, volumePatch)).Value;
             if (Mode != RecordedTestMode.Playback)
             {
@@ -413,12 +424,14 @@ namespace Azure.ResourceManager.NetApp.Tests
 
             //create Backup
             var backupName = Recording.GenerateAssetName("backup-");
-            NetAppBackupData backupData = new(DefaultLocation);
-            backupData.Label = "adHocBackup";
+            NetAppBackupData backupData = new(_defaultLocation)
+            {
+                Label = "adHocBackup"
+            };
             NetAppVolumeBackupResource backupResource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, backupName, backupData)).Value;
             Assert.IsNotNull(backupResource1);
             Assert.AreEqual(backupName, backupResource1.Id.Name);
-            NetAppVolumeBackupResource backupResource2 = await _volumeBackupCollection.GetAsync(backupName);
+            //NetAppVolumeBackupResource backupResource2 = await _volumeBackupCollection.GetAsync(backupName);
             if (Mode != RecordedTestMode.Playback)
             {
                 await Task.Delay(60000);
@@ -427,13 +440,15 @@ namespace Azure.ResourceManager.NetApp.Tests
 
             //create second Backup
             var backupName2 = Recording.GenerateAssetName("backup-");
-            NetAppBackupData backupData2 = new(DefaultLocation);
-            backupData2.Label = "adHocBackup2";
+            NetAppBackupData backupData2 = new(_defaultLocation)
+            {
+                Label = "adHocBackup2"
+            };
             NetAppVolumeBackupResource backup2Resource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, backupName2, backupData2)).Value;
             Assert.IsNotNull(backup2Resource1);
             Assert.AreEqual(backupName2, backup2Resource1.Id.Name);
             Assert.AreEqual(backupData2.Label, backup2Resource1.Data.Label);
-            NetAppVolumeBackupResource backup2Resource2 = await _volumeBackupCollection.GetAsync(backupName2);
+            //NetAppVolumeBackupResource backup2Resource2 = await _volumeBackupCollection.GetAsync(backupName2);
             if (Mode != RecordedTestMode.Playback)
             {
                 await Task.Delay(60000);
@@ -466,16 +481,16 @@ namespace Azure.ResourceManager.NetApp.Tests
         [RecordedTest]
         public async Task GetBackupStatus()
         {
-            //getVault id
-            _vaults.Should().HaveCount(1);
-            Assert.IsNotNull(_vault);
-
             //Update volume to enable backups
-            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, _vault.Id, true);
-            NetAppVolumePatchDataProtection dataProtectionProperties = new();
-            dataProtectionProperties.Backup = backupPolicyProperties;
-            NetAppVolumePatch volumePatch = new(DefaultLocation);
-            volumePatch.DataProtection = dataProtectionProperties;
+            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, true);
+            NetAppVolumePatchDataProtection dataProtectionProperties = new()
+            {
+                Backup = backupPolicyProperties
+            };
+            NetAppVolumePatch volumePatch = new(_defaultLocation)
+            {
+                DataProtection = dataProtectionProperties
+            };
             NetAppVolumeResource volumeResource1 = (await _volumeResource.UpdateAsync(WaitUntil.Completed, volumePatch)).Value;
             if (Mode != RecordedTestMode.Playback)
             {
@@ -491,8 +506,10 @@ namespace Azure.ResourceManager.NetApp.Tests
 
             //create Backup
             var backupName = Recording.GenerateAssetName("backup-");
-            NetAppBackupData backupData = new(DefaultLocation);
-            backupData.Label = "adHocBackup";
+            NetAppBackupData backupData = new(_defaultLocation)
+            {
+                Label = "adHocBackup"
+            };
             NetAppVolumeBackupResource backupResource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, backupName, backupData)).Value;
             Assert.IsNotNull(backupResource1);
             Assert.AreEqual(backupName, backupResource1.Id.Name);
@@ -520,16 +537,17 @@ namespace Azure.ResourceManager.NetApp.Tests
         [RecordedTest]
         public async Task CreateVolumeFromBackupCheckRestoreStatus()
         {
-            //getVault id
-            _vaults.Should().HaveCount(1);
-            Assert.IsNotNull(_vault);
             await WaitForVolumeSucceeded(_volumeCollection, _volumeResource);
             //Update volume to enable backups
-            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, _vault.Id, true);
-            NetAppVolumePatchDataProtection dataProtectionProperties = new();
-            dataProtectionProperties.Backup = backupPolicyProperties;
-            NetAppVolumePatch volumePatch = new(DefaultLocation);
-            volumePatch.DataProtection = dataProtectionProperties;
+            NetAppVolumeBackupConfiguration backupPolicyProperties = new(null, false, true);
+            NetAppVolumePatchDataProtection dataProtectionProperties = new()
+            {
+                Backup = backupPolicyProperties
+            };
+            NetAppVolumePatch volumePatch = new(_defaultLocation)
+            {
+                DataProtection = dataProtectionProperties
+            };
             NetAppVolumeResource volumeResource1 = (await _volumeResource.UpdateAsync(WaitUntil.Completed, volumePatch)).Value;
             if (Mode != RecordedTestMode.Playback)
             {
@@ -547,8 +565,10 @@ namespace Azure.ResourceManager.NetApp.Tests
 
             //create Backup
             var backupName = Recording.GenerateAssetName("backup-");
-            NetAppBackupData backupData = new(DefaultLocation);
-            backupData.Label = "adHocBackup";
+            NetAppBackupData backupData = new(_defaultLocation)
+            {
+                Label = "adHocBackup"
+            };
             NetAppVolumeBackupResource backupResource1 = (await _volumeBackupCollection.CreateOrUpdateAsync(WaitUntil.Completed, backupName, backupData)).Value;
             Assert.IsNotNull(backupResource1);
             Assert.AreEqual(backupName, backupResource1.Id.Name);
@@ -570,7 +590,7 @@ namespace Azure.ResourceManager.NetApp.Tests
             //Restore backup
             //You can restore a backup only to a new volume. You cannot overwrite the existing volume with the backup
             string newVolumeName = Recording.GenerateAssetName("restoredVolume-");
-            NetAppVolumeResource _restoredVolumeResource = await CreateVolume(DefaultLocation, NetAppFileServiceLevel.Premium, _defaultUsageThreshold, volumeName: newVolumeName, subnetId: DefaultSubnetId, backupId: backupResource2.Id);
+            NetAppVolumeResource _restoredVolumeResource = await CreateVolume(_defaultLocation, NetAppFileServiceLevel.Premium, _defaultUsageThreshold, volumeName: newVolumeName, subnetId: DefaultSubnetId, backupId: backupResource2.Id);
             if (Mode != RecordedTestMode.Playback)
             {
                 await Task.Delay(40000);
@@ -646,66 +666,9 @@ namespace Azure.ResourceManager.NetApp.Tests
             }
         }
 
-        private async Task WaitForRestoreStatusSucceeded(NetAppVolumeBackupCollection volumeBackupCollection, string backupName, NetAppVolumeResource volumeResource = null)
-        {
-            if (volumeResource == null)
-            {
-                volumeResource = _volumeResource;
-            }
-            var maxDelay = TimeSpan.FromSeconds(120);
-            int count = 0;
-            if (Environment.GetEnvironmentVariable("AZURE_TEST_MODE") == "Playback")
-            {
-                maxDelay = TimeSpan.FromMilliseconds(500);
-            }
-
-            IEnumerable<TimeSpan> delay = Backoff.DecorrelatedJitterBackoffV2(medianFirstRetryDelay: TimeSpan.FromSeconds(5), retryCount: 500)
-                    .Select(s => TimeSpan.FromTicks(Math.Min(s.Ticks, maxDelay.Ticks))); // use jitter strategy in the retry algorithm to prevent retries bunching into further spikes of load, with ceiling on delays (for larger retrycount)
-
-            Polly.Retry.AsyncRetryPolicy<bool> retryPolicy = Policy
-                .HandleResult<bool>(false) // retry if delegate executed asynchronously returns false
-                .WaitAndRetryAsync(delay);
-
-            try
-            {
-                await retryPolicy.ExecuteAsync(async () =>
-                {
-                    count++;
-                    NetAppVolumeBackupResource backup = await volumeBackupCollection.GetAsync(backupName);
-                    Console.WriteLine($"Get provisioning state for backup {backupName} run {count} provisioning state is {backup.Data.ProvisioningState}");
-                    if (backup.Data.ProvisioningState.Equals("Succeeded") || backup.Data.ProvisioningState.Equals("Failed"))
-                    {
-                        //Check status as well
-                        NetAppRestoreStatus restoreStatus = (await volumeResource.GetRestoreStatusAsync()).Value;
-                        Console.WriteLine($"Get RestoreStatus state volume: {volumeResource} run {count} BackupStatus.MirrorState {restoreStatus.MirrorState}, BackupStatus.RelationsShipt status {restoreStatus.RelationshipStatus}");
-                        if (restoreStatus.MirrorState == NetAppMirrorState.Mirrored)
-                        {
-                            return true;
-                        }
-                        else
-                        {
-                            return false;
-                        }
-                    }
-                    else
-                    {
-                        return false;
-                    }
-                }
-                );
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Final Throw {ex.Message}");
-                throw;
-            }
-        }
         private async Task WaitForVolumeSucceeded(NetAppVolumeCollection volumeCollection, NetAppVolumeResource volumeResource = null)
         {
-            if (volumeResource == null)
-            {
-                volumeResource = _volumeResource;
-            }
+            volumeResource ??= _volumeResource;
 
             var maxDelay = TimeSpan.FromSeconds(120);
             int count = 0;